despesas-app/src/main/java/br/com/despesas/service/DespesaService.java
package br.com.despesas.service;

import br.com.despesas.model.Despesa;
import br.com.despesas.repository.DespesaRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
public class DespesaService {

    @Autowired
    private DespesaRepository despesaRepository;

    @Transactional
    public Despesa salvar(Despesa despesa) {
        return despesaRepository.save(despesa);
    }

    public List<Despesa> buscarTodos() {
        return despesaRepository.findAll();
    }

}
